split app into smaller module files
