+++ lib/app-manager.js
@@ -1,97 +1,60 @@
 // @flow
 
-import { delay, retry } from './utils/timers';
 import AppManagerError from './utils/app-manager-error';
-import { setNames, defaultGetAppNameFromUrl } from './utils/config';
+import { setNames, defaultGetAppNameFromUrl, getErrorStatusDetails, getLoadingStatusDetails, getDefaultStatusDetails } from './utils/config';
+import shimHistory from './utils/shim-history';
 import initGetPageFragments from './utils/get-page-fragments';
+import initGetLifecycle from './utils/lifecycle';
+import initGetState from './utils/get-state';
 
 import type {
-  ScriptVersion3Type,
-  ScriptVersion4Type,
-  StatusDetailsType,
   EventsType,
-  ScriptType,
-  OptionsType,
+  OverrideOptionsType,
   ConfigType,
-  AppType,
   FragmentNameType,
   SlotNameType,
   StateType,
   StateChangeType,
+  LifecycleType,
 } from './index';
 
 import type Window from './utils/window-stub';
 
-import { DOM_CHECK_INTERVAL, DEFAULT_TIMEOUT, statuses, eventTitles, levels } from './constants';
-
-type LifecycleType = AsyncGenerator<boolean, boolean, boolean>;
-
-function getDefaultStatusDetails() {
-  return {
-    level: levels.INFO,
-    status: statuses.DEFAULT,
-  };
-}
-
-function getLoadingStatusDetails() {
-  return {
-    level: levels.INFO,
-    status: statuses.LOADING,
-  };
-}
-
-function getErrorStatusDetails(amError: AppManagerError) {
-  return {
-    level: amError.level || levels.ERROR,
-    status: statuses.ERROR,
-    message: amError.message,
-  };
-}
+import { DEFAULT_TIMEOUT, statuses, eventTitles, levels } from './constants';
 
 export default function initAppManager(w: Window) {
-  function getBrowserLocation(): { pathname: string, search: string } {
-    const errData = {
-      source: 'get_browser_location',
+  if (!w.document || !w.location || !w.history) {
+    throw new AppManagerError('Invalid window object. Aborting.', {
+      source: 'init',
+      level: levels.ERROR,
       fragment: null,
       slot: null,
       app: null,
       path: null,
-    };
-
-    const { location } = w;
-
-    if (!location || typeof location.pathname !== 'string') {
-      throw new AppManagerError('Could not get location from browser.', Object.assign({}, errData, {
-        level: levels.ERROR,
-        code: 'no_location',
-        recoverable: false,
-      }));
-    }
-
-    return location;
+      code: 'invalid_environment',
+      recoverable: false,
+    });
   }
 
-  function selectFromDom(querySelector: string): ?Element {
-    return w.document.querySelector(querySelector);
-  }
-
-  function appManager(config: ConfigType, events: EventsType, overrideOptions?: OptionsType) {
-    let runningStateChange: ?StateChangeType = null;
-    let queuedStateChange: ?() => StateChangeType = null;
-    let state: StateType;
-
-    if (!w.document || !w.location || !w.history) {
-      throw new AppManagerError('Invalid window object. Aborting.', {
-        source: 'constructor',
+  function appManager(config: ConfigType, events: EventsType, overrideOptions?: OverrideOptionsType) {
+    if (!events || typeof events.emit !== 'function' || typeof events.on !== 'function' || typeof events.removeListener !== 'function') {
+      throw new AppManagerError('Invalid events object. Aborting.', {
+        source: 'init',
         level: levels.ERROR,
         fragment: null,
         slot: null,
         app: null,
         path: null,
-        code: 'invalid_environment',
+        code: 'invalid_events',
         recoverable: false,
       });
     }
+
+    let runningStateChange: ?StateChangeType = null;
+    let state: StateType;
+
+    const mountedLifecycles: { [slotName: SlotNameType]: LifecycleType } = {};
+    let queuedStateChange: ?() => StateChangeType = null;
 
     const apps = setNames(config.apps);
     const slots = setNames(config.slots);
@@ -104,419 +67,10 @@ export default function initAppManager(w: Window) {
     };
 
     const options = Object.assign({}, defaultOptions, overrideOptions || {});
+
+    const getState = initGetState(w, options, apps);
     const getPageFragments = initGetPageFragments(apps, fragments, slots);
-
-    const cachedScripts: { [fragmentName: FragmentNameType]: ScriptType } = {};
-    const mountedLifecycles: { [slotName: SlotNameType]: LifecycleType } = {};
-
-    async function getState(prevApp: ?AppType, eventTitle?: mixed): Promise<?StateType> {
-      const { pathname: path, search } = getBrowserLocation();
-      const url = `${path}${search}`;
-
-      const errData = {
-        source: 'set_state',
-        fragment: null,
-        slot: null,
-        path,
-      };
-
-      let appName = null;
-
-      if (typeof options.getAppNameFromUrl === 'function') {
-        try {
-          appName = await options.getAppNameFromUrl(url);
-        } catch (err) {
-          throw new AppManagerError(err, Object.assign({}, errData, {
-            app: appName,
-            level: levels.ERROR,
-            code: 'get_app_name_from_location',
-            recoverable: false,
-          }));
-        }
-      }
-
-      if (!appName) {
-        return null;
-      }
-
-      let additionalState = {};
-
-      if (typeof options.getAdditionalState === 'function') {
-        try {
-          additionalState = await options.getAdditionalState(appName, url);
-        } catch (err) {
-          throw new AppManagerError(err, Object.assign({}, errData, {
-            app: appName,
-            level: levels.ERROR,
-            code: 'get_additional_state',
-            recoverable: false,
-          }));
-        }
-      }
-
-      return Object.assign({}, additionalState, {
-        event: eventTitle && typeof eventTitle === 'string' ? eventTitle : null,
-        path,
-        prevApp,
-        app: apps[appName],
-      });
-    }
-
-    function emit(eventTitle, data) {
-      if (events && typeof events.emit === 'function') {
-        events.emit(eventTitle, data);
-      }
-    }
-
-    function on(eventTitle, listener) {
-      if (events && typeof events.on === 'function') {
-        events.on(eventTitle, listener);
-      }
-    }
-
-    function removeListener(eventTitle, listener) {
-      if (events && typeof events.removeListener === 'function') {
-        events.removeListener(eventTitle, listener);
-      }
-    }
-
-    async function getSlotElement(slotName: SlotNameType): Promise<?Element> {
-      const errData = {
-        source: 'get_slot_element',
-        fragment: null,
-        slot: null,
-        app: state ? state.app.name : null,
-        path: state ? state.path : null,
-      };
-
-      const slot = slots[slotName];
-
-      const { querySelector } = slot;
-
-      if (!querySelector) {
-        return null;
-      }
-
-      const element = await retry(
-        () => selectFromDom(querySelector),
-        DOM_CHECK_INTERVAL,
-        options.importTimeout,
-      );
-
-      if (!element) {
-        throw new AppManagerError(`Element ${querySelector} not found for slot ${slot.name}.`, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'element_not_found',
-          recoverable: false,
-        }));
-      }
-
-      return element;
-    }
-
-    async function loadScript(fragmentName: FragmentNameType, slotName: SlotNameType): Promise<ScriptType> {
-      const errData = {
-        source: 'load_script',
-        fragment: fragmentName,
-        slot: slotName,
-        app: state ? state.app.name : null,
-        path: state ? state.path : null,
-      };
-
-      let script;
-
-      try {
-        script = await Promise.race([
-          fragments[fragmentName].loadScript(state),
-          delay(options.importTimeout),
-        ]);
-      } catch (err) {
-        throw new AppManagerError(err, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'load_script',
-          recoverable: true,
-        }));
-      }
-
-      if (!script) {
-        throw new AppManagerError(`loadScript timed out for fragment ${fragmentName}.`, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'time_out',
-          recoverable: true,
-        }));
-      }
-
-      if (!('version' in script)) {
-        throw new AppManagerError(`Loaded script for fragment ${fragmentName} does not have a version property.`, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'invalid_script',
-          recoverable: true,
-        }));
-      }
-
-      return script;
-    }
-
-    async function updateStatus(script: ScriptType, statusDetails: StatusDetailsType): Promise<void> {
-      if (script.version === 3) {
-        return;
-      }
-
-      try {
-        if (script.version === 4) {
-          if (typeof script.onUpdateStatus === 'function') {
-            await script.onUpdateStatus(statusDetails.status, apps[state.app.name]);
-          }
-        } else if (typeof script.onUpdateStatus === 'function') {
-          await script.onUpdateStatus(statusDetails, state);
-        }
-      } catch (err) {
-        // @TODO Should these be ignored?
-      }
-    }
-
-    async function hydrate(script: ScriptType, fragmentName: FragmentNameType, slotName: SlotNameType): Promise<void> {
-      const slotElement = await getSlotElement(slotName);
-
-      const errData = {
-        source: 'hydrate',
-        fragment: fragmentName,
-        slot: slotName,
-        app: state ? state.app.name : null,
-        path: state ? state.path : null,
-      };
-
-      if (!slotElement) {
-        throw new AppManagerError(`Could not get DOM element for fragment ${fragmentName}`, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'element_not_found',
-          recoverable: true,
-        }));
-      }
-
-      try {
-        if (script.version === 3) {
-          await (script: ScriptVersion3Type).hydrate(slotElement, w.history, state.app);
-        } else if (script.version === 4) {
-          if (typeof script.hydrate === 'function') {
-            await script.hydrate(slotElement, state.app);
-          } else {
-            await script.mount(slotElement, eventTitles.HISTORY_REPLACE_STATE, state.app);
-          }
-        } else if (typeof script.hydrate === 'function') {
-          await script.hydrate(slotElement, state);
-        } else if (typeof script.render === 'function') {
-          await script.render(slotElement, state);
-        }
-      } catch (err) {
-        throw new AppManagerError(err, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'hydrate',
-          recoverable: true,
-        }));
-      }
-    }
-
-    async function render(script: ScriptType, fragmentName: FragmentNameType, slotName: SlotNameType): Promise<void> {
-      const slotElement = await getSlotElement(slotName);
-
-      const errData = {
-        source: 'render',
-        fragment: fragmentName,
-        slot: slotName,
-        app: state ? state.app.name : null,
-        path: state ? state.path : null,
-      };
-
-      if (!slotElement) {
-        throw new AppManagerError(`Could not get DOM element for fragment ${fragmentName}`, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'element_not_found',
-          recoverable: true,
-        }));
-      }
-
-      try {
-        if (script.version === 3) {
-          await (script: ScriptVersion3Type).mount(slotElement, w.history, state.app);
-        } else if (script.version === 4) {
-          if (state.event) {
-            await (script: ScriptVersion4Type).mount(slotElement, state.event, state.app);
-          }
-        } else if (typeof script.render === 'function') {
-          await script.render(slotElement, state);
-        }
-      } catch (err) {
-        throw new AppManagerError(err, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'render',
-          recoverable: true,
-        }));
-      }
-    }
-
-    async function onStateChange(script: ScriptType, fragmentName: FragmentNameType, slotName: SlotNameType): Promise<void> {
-      const errData = {
-        source: 'on_state_change',
-        fragment: fragmentName,
-        slot: slotName,
-        app: state ? state.app.name : null,
-        path: state ? state.path : null,
-      };
-
-      try {
-        if (script.version === 3) {
-          await script.onStateChange(w.history, state.app);
-        } else if (script.version === 4) {
-          if (typeof script.onStateChange === 'function') {
-            if (state.event) {
-              await script.onStateChange(state.event, state.app);
-            }
-          }
-        } else if (typeof script.onStateChange === 'function') {
-          await script.onStateChange(state);
-        }
-      } catch (err) {
-        throw new AppManagerError(err, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'on_state_change',
-          recoverable: true,
-        }));
-      }
-    }
-
-    async function unmount(script: ScriptType, fragmentName: FragmentNameType, slotName: SlotNameType): Promise<void> {
-      const errData = {
-        source: 'unmount',
-        fragment: fragmentName,
-        slot: slotName,
-        app: state ? state.app.name : null,
-        path: state ? state.path : null,
-      };
-
-      const slotElement = await getSlotElement(slotName);
-
-      if (!slotElement) {
-        throw new AppManagerError(`Could not get DOM element for fragment ${fragmentName}`, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'element_not_found',
-          recoverable: true,
-        }));
-      }
-
-      if (!state.prevApp) {
-        throw new AppManagerError(`Invalid state at unmount for fragment ${fragmentName}`, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'invalid_state',
-          recoverable: true,
-        }));
-      }
-
-      try {
-        if (script.version === 3) {
-          await script.unmount(slotElement, w.history, state.prevApp);
-        } else if (script.version === 4) {
-          if (state.event) {
-            await script.unmount(slotElement, state.event, state.prevApp);
-          }
-        } else if (typeof script.unmount === 'function') {
-          await script.unmount(slotElement, state);
-        }
-      } catch (err) {
-        throw new AppManagerError(err, Object.assign({}, errData, {
-          level: levels.ERROR,
-          code: 'unmount',
-          recoverable: true,
-        }));
-      }
-    }
-
-    async function* scriptLifecycle(fragmentName: FragmentNameType, slotName: SlotNameType, initialRender: boolean): LifecycleType {
-      let statechange;
-
-      const loadingStatusDetails = getLoadingStatusDetails();
-      const defaultStatusDetails = getDefaultStatusDetails();
-
-      const cachedScript = cachedScripts[fragmentName];
-
-      let script;
-
-      if (cachedScript) {
-        script = cachedScript;
-      } else {
-        script = await loadScript(fragmentName, slotName);
-
-        cachedScripts[fragmentName] = script;
-      }
-
-      if (!script) {
-        return false;
-      }
-
-      while (!statechange) {
-        try {
-          statechange = yield true;
-          break;
-        } catch (err) {
-          /* eslint-disable no-await-in-loop */
-          await updateStatus(script, getErrorStatusDetails(err));
-        }
-      }
-
-      await updateStatus(script, loadingStatusDetails);
-
-      if (initialRender) {
-        await hydrate(script, fragmentName, slotName);
-      } else {
-        await render(script, fragmentName, slotName);
-      }
-
-      await updateStatus(script, defaultStatusDetails);
-
-      while (!statechange) {
-        try {
-          statechange = yield true;
-          break;
-        } catch (err) {
-          /* eslint-disable no-await-in-loop */
-          await updateStatus(script, getErrorStatusDetails(err));
-        }
-      }
-
-      while (!statechange) {
-        await updateStatus(script, loadingStatusDetails);
-        await onStateChange(script, fragmentName, slotName);
-        await updateStatus(script, defaultStatusDetails);
-
-        while (true) {
-          try {
-            statechange = yield true;
-            break;
-          } catch (err) {
-            /* eslint-disable no-await-in-loop */
-            await updateStatus(script, getErrorStatusDetails(err));
-          }
-        }
-      }
-
-      yield true;
-
-      await updateStatus(script, loadingStatusDetails);
-      await unmount(script, fragmentName, slotName);
-      await updateStatus(script, defaultStatusDetails);
-
-      while (!statechange) {
-        try {
-          statechange = yield true;
-          break;
-        } catch (err) {
-          /* eslint-disable no-await-in-loop */
-          await updateStatus(script, getErrorStatusDetails(err));
-        }
-      }
-
-      return false;
-    }
+    const getLifecycle = initGetLifecycle(w, options, apps, fragments, slots);
 
     async function handleError(slotName: SlotNameType, err: AppManagerError): Promise<void> {
       try {
@@ -525,14 +79,14 @@ export default function initAppManager(w: Window) {
       } catch (_) {
         // @TODO Script has already thrown
       } finally {
-        emit(eventTitles.ERROR, err);
-        emit(eventTitles.STATUS_CHANGE, getErrorStatusDetails(err));
+        events.emit(eventTitles.ERROR, err);
+        events.emit(eventTitles.STATUS_CHANGE, getErrorStatusDetails(err));
       }
     }
 
     async function nextAction(slotName: SlotNameType): Promise<boolean> {
       try {
-        const { value } = await mountedLifecycles[slotName].next(true);
+        const { value } = await mountedLifecycles[slotName].next(state);
         return value === true;
       } catch (err) {
         await handleError(slotName, err);
@@ -562,7 +116,7 @@ export default function initAppManager(w: Window) {
           return false;
         }
 
-        mountedLifecycles[slotName] = scriptLifecycle(fragmentName, slotName, initialRender);
+        mountedLifecycles[slotName] = getLifecycle(state, fragmentName, slotName, initialRender);
       } else {
         delete mountedLifecycles[slotName];
       }
@@ -579,11 +133,11 @@ export default function initAppManager(w: Window) {
         path: state ? state.path : null,
       };
 
-      await mountedLifecycles[slotName].next(true);
+      await mountedLifecycles[slotName].next(state);
 
       await nextAction(slotName);
 
-      const { done } = await mountedLifecycles[slotName].next(true);
+      const { done } = await mountedLifecycles[slotName].next(state);
 
       if (!done) {
         await handleError(slotName, new AppManagerError('Unmounting a fragment that is not finished.', Object.assign({}, errData, {
@@ -655,7 +209,7 @@ export default function initAppManager(w: Window) {
         });
       }
 
-      emit(eventTitles.STATUS_CHANGE, getLoadingStatusDetails());
+      events.emit(eventTitles.STATUS_CHANGE, getLoadingStatusDetails());
 
       await Promise.all(slotsToUnmount.map((slotName) => unmountSlot(slotName)));
 
@@ -666,10 +220,10 @@ export default function initAppManager(w: Window) {
       if (haveScripts.every(Boolean)) {
         await Promise.all(slotsToMount.map((slotName) => nextAction(slotName)));
       } else {
-        emit(eventTitles.EXTERNAL_LINK);
+        events.emit(eventTitles.EXTERNAL_LINK);
       }
 
-      emit(eventTitles.STATUS_CHANGE, getDefaultStatusDetails());
+      events.emit(eventTitles.STATUS_CHANGE, getDefaultStatusDetails());
     }
 
     function getStateChanger(initialRender: boolean, eventTitle?: mixed) {
@@ -679,21 +233,21 @@ export default function initAppManager(w: Window) {
           const newState = await getState(currentApp, eventTitle);
 
           if (!newState) {
-            emit(eventTitles.EXTERNAL_LINK);
+            events.emit(eventTitles.EXTERNAL_LINK);
             return;
           }
 
           state = newState;
         } catch (err) {
-          emit(eventTitles.ERROR, err);
+          events.emit(eventTitles.ERROR, err);
           return;
         }
 
         try {
           await updatePage(initialRender);
         } catch (err) {
-          emit(eventTitles.ERROR, err);
-          emit(eventTitles.STATUS_CHANGE, getErrorStatusDetails(err));
+          events.emit(eventTitles.ERROR, err);
+          events.emit(eventTitles.STATUS_CHANGE, getErrorStatusDetails(err));
         }
 
         const nextStateChange = queuedStateChange;
@@ -717,61 +271,7 @@ export default function initAppManager(w: Window) {
       }
     }
 
-    (function shimHistory() {
-      /* eslint-disable no-param-reassign */
-
-      const { onpopstate, onbeforeunload } = w;
-      const { pushState, replaceState } = w.history;
-
-      w.history.pushState = (...args) => {
-        if (typeof pushState === 'function') {
-          pushState.apply(w.history, args);
-        }
-
-        emit(eventTitles.HISTORY_PUSH_STATE);
-        emit(eventTitles.HISTORY_STATE_CHANGE, eventTitles.HISTORY_PUSH_STATE);
-      };
-
-      w.history.replaceState = (...args) => {
-        if (typeof replaceState === 'function') {
-          replaceState.apply(w.history, args);
-        }
-
-        emit(eventTitles.HISTORY_REPLACE_STATE);
-        emit(eventTitles.HISTORY_STATE_CHANGE, eventTitles.HISTORY_REPLACE_STATE);
-      };
-
-      // $FlowFixMe
-      w.history.getState = () => ({
-        data: w.history.state,
-        title: w.document.title,
-        url: w.location.href,
-        hash: `${w.location.pathname}${w.location.search || ''}`,
-      });
-
-      w.onpopstate = (...args) => {
-        if (typeof onpopstate === 'function') {
-          onpopstate.apply(w.history, args);
-        }
-
-        emit(eventTitles.HISTORY_POP_STATE);
-        emit(eventTitles.HISTORY_STATE_CHANGE, eventTitles.HISTORY_POP_STATE);
-      };
-
-      w.onbeforeunload = (...args) => {
-        if (typeof onbeforeunload === 'function') {
-          onbeforeunload.apply(w.history, args);
-        }
-
-        removeListener(eventTitles.HISTORY_STATE_CHANGE, handleStateChange);
-
-        w.onpopstate = onpopstate;
-        w.history.pushState = pushState;
-        w.history.replaceState = replaceState;
-      };
-
-      on(eventTitles.HISTORY_STATE_CHANGE, handleStateChange);
-    }());
+    shimHistory(w, events, handleStateChange);
 
     runningStateChange = getStateChanger(true)();
