+++ lib/index.js.flow
@@ -10,6 +10,13 @@ export type StatusType = $Values<typeof statuses>;
 export type StatusLevelType = $Values<typeof levels>;
 
 export type OptionsType = {
+  importTimeout: number,
+  haltOnError: boolean,
+  getAppNameFromUrl: (url: string) => ?AppNameType,
+  getAdditionalState?: (appName: AppNameType, url: string) => { [stateName: string]: mixed },
+};
+
+export type OverrideOptionsType = {
   importTimeout?: number,
   haltOnError?: boolean,
   getAppNameFromUrl?: (url: string) => ?AppNameType,
@@ -24,6 +31,8 @@ type BaseAppType = {
 }
 
 export type AppType = BaseAppType & { name: AppNameType };
+
+export type AppsMapType = { [appName: AppNameType]: AppType };
 
 export type EventCallbackType = (...data: Array<mixed>) => mixed;
 export type EmitterType = (eventTitle: EventTitleType, ...data: Array<mixed>) => void;
@@ -82,6 +91,8 @@ type BaseSlotType = { querySelector: ?string }
 
 export type SlotType = BaseSlotType & { name: SlotNameType };
 
+export type SlotsMapType = { [slotName: SlotNameType]: SlotType };
+
 type LoadScriptType = (state: StateType) => Promise<ScriptType>;
 
 type GetMarkupType = (state: StateType, ...otherArgs: Array<mixed>) => Promise<string>;
@@ -97,11 +108,15 @@ export type StateChangeType = Promise<void>;
 
 export type FragmentType = BaseFragmentType & { name: FragmentNameType };
 
+export type FragmentsMapType = { [fragmentName: FragmentNameType]: FragmentType };
+
 export type ConfigType = {|
   apps: { [appName: AppNameType]: BaseAppType },
   slots: { [slotName: SlotNameType]: BaseSlotType },
   fragments: { [fragmentName: FragmentNameType]: BaseFragmentType },
 |};
+
+export type LifecycleType = AsyncGenerator<boolean, boolean, StateType>;
 
 declare export default (config: ConfigType, events: EventsType, options?: OptionsType) => {
   getState(): StateType,
