+++ lib/server.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { ConfigType, OptionsType } from './index';
+import type { ConfigType, OptionsType, SlotNameType } from './index';
 
 import AppManagerError from './utils/app-manager-error';
 import { setNames, defaultGetAppNameFromUrl } from './utils/config';
@@ -20,7 +20,7 @@ export default function appManager(config: ConfigType, overrideOptions?: Options
   const getPageFragments = initGetPageFragments(apps, fragments, slots);
 
   return {
-    async getSlotsMarkup(url: string, ...getMarkupArgs: Array<mixed>): Promise<{ [slotName: string]: string }> {
+    async getSlotsMarkup(url: string, ...getMarkupArgs: Array<mixed>): Promise<{ [slotName: SlotNameType]: string }> {
       let currentAppName = null;
       const [path] = url.split('?');
       const errData = {
