+++ lib/utils/get-page-fragments.js
@@ -6,9 +6,9 @@ import type {
   AppNameType,
   FragmentNameType,
   SlotNameType,
-  AppType,
-  FragmentType,
-  SlotType,
+  AppsMapType,
+  FragmentsMapType,
+  SlotsMapType,
 } from '../index';
 
 import AppManagerError from './app-manager-error';
@@ -16,7 +16,7 @@ import { levels } from '../constants';
 
 type PageFragmentsType = { [slotName: SlotNameType]: ?FragmentNameType }
 
-export default function initGetPageFragments(apps: { [appName: AppNameType]: AppType }, fragments: { [fragmentName: FragmentNameType]: FragmentType }, slots: { [slotName: SlotNameType]: SlotType }) {
+export default function initGetPageFragments(apps: AppsMapType, fragments: FragmentsMapType, slots: SlotsMapType) {
   return function getPageFragments(appName: AppNameType, path: ?string): PageFragmentsType {
     const errData = {
       source: 'get_page_fragments',
